Make DeleteSecurityGroupStage and DeleteSecurityGroupTask generic.
